// react/darshana-refactor-web/src/interfaces/projectAplication.ts
// Generated by https://quicktype.io

export interface IProjectUser {
  status: boolean;
  message: string;
  data: Datum[];
}
export interface IProjectIndidual {
  project: Project;
  id: number;
  user_uuid: string;
  project_id: number;
  proposal: string;
  procedure_text: string;
  accept_price: boolean;
  price_proposal: null;
  accept_time: boolean;
  days: number;
  ready_to_close: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: null;
  accepted: number;
  close_at_proposal: string;
  recieve_pay_at: null | string;
  send_pay_at: null;
  recieve_pay_order_id: null;
  send_pay_order_id: null;
  talent_payment: null;
  weeks: number;
  shown_accepted_message: boolean;
  algorand_transaction: null | string;
  near_transaction: null | string;
}
interface Datum {
  time_counter_proposal: number;
  id: number;
  user_uuid: string;
  project_id: number;
  proposal: string;
  procedure_text: string;
  accept_price: boolean;
  price_proposal: null;
  accept_time: boolean;
  days: number;
  ready_to_close: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: null;
  accepted: number;
  close_at_proposal: string;
  recieve_pay_at: null;
  send_pay_at: null;
  recieve_pay_order_id: null;
  send_pay_order_id: null;
  talent_payment: null;
  weeks: number;
  project: Project;
  topic: string | null;
  algorand_transaction: null | string;
  near_transaction: null | string;
  counter_proposal_status: number;
  salary_counter_proposal: number;
}
interface ProjectInvidual {
  id: number;
  recruiter_id: null;
  business_id: number;
  project_status_id: number;
  topic_id: number;
  name: string;
  description: string;
  status: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: null;
  body: string;
  price: number;
  image_url: string;
  days: number;
  end_date: string;
  expected_close_at: null;
  weeks: number;
  is_visible: boolean;
  user_uuid: string;
  work_modality_id: number;
  country_id: number;
}
interface Project {
  id: number;
  recruiter_id: null;
  business_id: number;
  project_status_id: number;
  topic_id: number;
  name: string;
  description: string;
  status: boolean;
  created_at: string;
  updated_at: string;
  deleted_at: null;
  body: string;
  price: number;
  image_url: string;
  days: number;
  end_date: string;
  expected_close_at: null;
  weeks: number;
  is_visible: boolean;
  user_uuid: string;
  work_modality_id: number;
  country_id: number;
  country: Country;
  business: Business;
}

interface Business {
  id: number;
  user_workplace_id: number;
  user_uuid: string;
  created_at: string;
  updated_at: string;
  deleted_at: null;
  workplace: Country;
}

interface Country {}